Option to control pretty-print output. (#34)
* Pretty print behavior can now be controlled through a command line
option:
  * `--color never` disables text stylisation completely
* `--color auto` enables text stylisation only when the standard output
is connected to a terminal (i.e. not redirected to a file)
  * `--color always` forces text stylisation regardless of the output
* Colors and styles can be overriden using `--color-overrides` command
line option. The value is expected to be a string in [the `grep`
format](https://man7.org/linux/man-pages/man1/grep.1.html#ENVIRONMENT).
* Renamed `FormattingOptions` to `OutputOptions`. `FormattingOptions` is
now used as a name for a new `struct` that holds only pretty-print
options.
* Changed the underlying terminal coloring library from `colored` to
`yansi`.
* Consecutive characters are now painted as a whole (opposed to being
painted individually before). This should make no visual difference but
allows for easier validation of the results and more intuitive behavior.